backend/requirements.txt

flask
delta-spark==2.4.0
pyspark
pandas
flask-cors